fix(messaging): address code review P1/P2 findings
- PostgreSqlStorageInitializer: use DateTimeKind.Utc for DateTime.MinValue
  on TIMESTAMPTZ column — Npgsql 7+ rejects Kind=Unspecified
- SqlServerDataStorage: cast TtlSeconds to int for DATEADD — SQL Server
  rejects double/float for integer datepart argument
- PostgreSqlMonitoringApi._GetTimelineStats: GetInt64 for COUNT() bigint
  return type (same fix already applied to GetStatisticsAsync)
- PostgreSqlMonitoringApi: forward CancellationToken to _GetNumberOfMessage
  from all four count methods
- Remove unnecessary await Task.CompletedTask in test

+++ src/Headless.Messaging.SqlServer/SqlServerDataStorage.cs
@@ -103,7 +103,7 @@ public sealed class SqlServerDataStorage(
         [
             new SqlParameter("@Key", key),
             new SqlParameter("@Instance", instance),
-            new SqlParameter("@TtlSeconds", ttl.TotalSeconds),
+            new SqlParameter("@TtlSeconds", (int)ttl.TotalSeconds),
         ];
         await connection
             .ExecuteNonQueryAsync(sql, cancellationToken: cancellationToken, sqlParams: sqlParams)
